Whole code refactor, split service layer to service interface and service implementation
