+++ src/main/java/pl/medos/cmmsApi/CmmsApiApplication.java
@@ -2,13 +2,12 @@ package pl.medos.cmmsApi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
 @SpringBootApplication
 public class CmmsApiApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CmmsApiApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CmmsApiApplication.class, args);
+    }
 
 }
