+++ src/main/java/pl/medos/cmmsApi/controllers/WebDepartmentController.java
@@ -4,9 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 import pl.medos.cmmsApi.model.Department;
-import pl.medos.cmmsApi.model.Machine;
 import pl.medos.cmmsApi.service.DepartmentService;
-import pl.medos.cmmsApi.service.MachineService;
 
 import java.util.List;
 import java.util.logging.Logger;
@@ -25,7 +23,7 @@ public class WebDepartmentController {
     @GetMapping
     public String listView(ModelMap modelMap) {
         LOGGER.info("listView()");
-        List<Department> departments = departmentService.list();
+        List<Department> departments = departmentService.findAllDepartments();
         modelMap.addAttribute("departments", departments);
         LOGGER.info("listView(...)" + departments);
         return "list-department.html";
@@ -36,7 +34,7 @@ public class WebDepartmentController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws Exception {
         LOGGER.info("updateView()");
-        Department department = departmentService.read(id);
+        Department department = departmentService.findDepartmentById(id);
         modelMap.addAttribute("department", department);
         return "update-department.html";
     }
@@ -45,7 +43,7 @@ public class WebDepartmentController {
     public String update(
             @ModelAttribute(name = "department") Department department) {
         LOGGER.info("update()" + department);
-        Department updatedDepartment = departmentService.update(department);
+        Department updatedDepartment = departmentService.updateDepartment(department);
         LOGGER.info("update(...)" + updatedDepartment);
         return "redirect:/departments";
     }
@@ -54,16 +52,16 @@ public class WebDepartmentController {
     public String createView(ModelMap modelMap) {
         LOGGER.info("createView()");
         modelMap.addAttribute("department", new Department());
-        List<Department> departments = departmentService.list();
+        List<Department> departments = departmentService.findAllDepartments();
         modelMap.addAttribute("departments", departments);
         return "create-department.html";
     }
 
     @PostMapping(value = "/create")
     public String create(
-                   @ModelAttribute(name = "department") Department department) {
-        LOGGER.info("create(" + department+ ")");
-        Department savedDepartment = departmentService.create(department);
+            @ModelAttribute(name = "department") Department department) {
+        LOGGER.info("create(" + department + ")");
+        Department savedDepartment = departmentService.createDepartment(department);
         LOGGER.info("create(...)" + savedDepartment);
         return "redirect:/departments";
     }
@@ -73,7 +71,7 @@ public class WebDepartmentController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws Exception {
         LOGGER.info("read(" + id + ")");
-        Department department = departmentService.read(id);
+        Department department = departmentService.findDepartmentById(id);
         modelMap.addAttribute("department", department);
         return "read-department.html";
     }
@@ -82,7 +80,7 @@ public class WebDepartmentController {
     public String delete(
             @PathVariable(name = "id") Long id) {
         LOGGER.info("delete()");
-       departmentService.delete(id);
+        departmentService.deleteDepartment(id);
         return "redirect:/departments";
     }
 }
