+++ src/main/java/pl/medos/cmmsApi/controllers/WebInvoiceController.java
@@ -3,13 +3,11 @@ package pl.medos.cmmsApi.controllers;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
-import pl.medos.cmmsApi.exception.DepartmentNotFoundException;
 import pl.medos.cmmsApi.exception.InvoiceNotFoundException;
-import pl.medos.cmmsApi.exception.SupplierNotFoundException;
 import pl.medos.cmmsApi.model.Invoice;
 import pl.medos.cmmsApi.model.Supplier;
-import pl.medos.cmmsApi.service.InvoiceService;
-import pl.medos.cmmsApi.service.SupplierService;
+import pl.medos.cmmsApi.service.impl.InvoiceServiceImpl;
+import pl.medos.cmmsApi.service.impl.SupplierServiceImpl;
 
 import java.util.List;
 import java.util.logging.Logger;
@@ -20,20 +18,20 @@ public class WebInvoiceController {
 
     private static final Logger LOGGER = Logger.getLogger(WebInvoiceController.class.getName());
 
-    private InvoiceService invoiceService;
-    private SupplierService supplierService;
+    private InvoiceServiceImpl invoiceServiceImpl;
+    private SupplierServiceImpl supplierServiceImpl;
 
-    public WebInvoiceController(InvoiceService invoiceService, SupplierService supplierService) {
-        this.invoiceService = invoiceService;
-        this.supplierService = supplierService;
+    public WebInvoiceController(InvoiceServiceImpl invoiceServiceImpl, SupplierServiceImpl supplierServiceImpl) {
+        this.invoiceServiceImpl = invoiceServiceImpl;
+        this.supplierServiceImpl = supplierServiceImpl;
     }
 
     @GetMapping
     public String listView(ModelMap modelMap) {
         LOGGER.info("listView()");
-        List<Invoice> invoices = invoiceService.list();
+        List<Invoice> invoices = invoiceServiceImpl.findAllInvoices();
         modelMap.addAttribute("invoices", invoices);
-        List<Supplier> suppliers = supplierService.list();
+        List<Supplier> suppliers = supplierServiceImpl.list();
         modelMap.addAttribute("supplier", suppliers);
         LOGGER.info("listView(...)");
         return "list-invoice.html";
@@ -43,7 +41,7 @@ public class WebInvoiceController {
     public String createView(ModelMap modelMap) {
         LOGGER.info("createView()");
         modelMap.addAttribute("invoice", new Invoice());
-        List<Supplier> suppliers = supplierService.list();
+        List<Supplier> suppliers = supplierServiceImpl.list();
         modelMap.addAttribute("suppliers", suppliers);
         LOGGER.info("createView(...)");
         return "create-invoice.html";
@@ -52,7 +50,7 @@ public class WebInvoiceController {
     @PostMapping(value = "/create")
     public String create(@ModelAttribute(name = "invoice") Invoice invoice) {
         LOGGER.info("create()");
-        Invoice savedInvoice = invoiceService.create(invoice);
+        Invoice savedInvoice = invoiceServiceImpl.createInvoice(invoice);
         LOGGER.info("create(...)" + savedInvoice);
         return "redirect:/invoices";
     }
@@ -62,7 +60,7 @@ public class WebInvoiceController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws InvoiceNotFoundException {
         LOGGER.info("read(" + id + ")");
-        Invoice invoiceModel = invoiceService.read(id);
+        Invoice invoiceModel = invoiceServiceImpl.findInvoiceById(id);
         modelMap.addAttribute("invoice", invoiceModel);
         LOGGER.info("read(...)" + invoiceModel);
         return "read-invoice.html";
@@ -73,9 +71,9 @@ public class WebInvoiceController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws InvoiceNotFoundException {
         LOGGER.info("update()" + id);
-        Invoice invoice = invoiceService.read(id);
+        Invoice invoice = invoiceServiceImpl.findInvoiceById(id);
         modelMap.addAttribute("invoice", invoice);
-        List<Supplier> suppliers = supplierService.list();
+        List<Supplier> suppliers = supplierServiceImpl.list();
         modelMap.addAttribute("suppliers", suppliers);
         LOGGER.info("update(...)");
         return "create-invoice.html";
@@ -85,7 +83,7 @@ public class WebInvoiceController {
     public String update(
             @ModelAttribute(name = "invoice") Invoice invoice) throws InvoiceNotFoundException {
         LOGGER.info("update()" + invoice);
-        Invoice updatedInvoice = invoiceService.update(invoice);
+        Invoice updatedInvoice = invoiceServiceImpl.updateInvoice(invoice);
         LOGGER.info("update(...)" + updatedInvoice);
         return "redirect:/invoices";
     }
@@ -93,7 +91,7 @@ public class WebInvoiceController {
     @GetMapping(value = "/delete/{id}")
     public String delete(@PathVariable(name = "id") Long id) {
         LOGGER.info("delete(" + id + ")");
-        invoiceService.delete(id);
+        invoiceServiceImpl.deleteInvoice(id);
         LOGGER.info("delete(...)");
         return "redirect:/invoices";
     }
