+++ src/main/java/pl/medos/cmmsApi/controllers/WebMachineController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import pl.medos.cmmsApi.model.Department;
 import pl.medos.cmmsApi.model.Machine;
 import pl.medos.cmmsApi.service.DepartmentService;
-import pl.medos.cmmsApi.service.MachineService;
+import pl.medos.cmmsApi.service.impl.MachineServiceImpl;
 
 import java.util.List;
 import java.util.logging.Logger;
@@ -17,18 +17,18 @@ public class WebMachineController {
 
     private static final Logger LOGGER = Logger.getLogger(WebMachineController.class.getName());
 
-    private MachineService machineService;
+    private MachineServiceImpl machineServiceImpl;
     private DepartmentService departmentService;
 
-    public WebMachineController(MachineService machineService, DepartmentService departmentService) {
-        this.machineService = machineService;
+    public WebMachineController(MachineServiceImpl machineServiceImpl, DepartmentService departmentService) {
+        this.machineServiceImpl = machineServiceImpl;
         this.departmentService = departmentService;
     }
 
     @GetMapping
     public String listView(ModelMap modelMap) {
         LOGGER.info("listView()");
-        List<Machine> machines = machineService.list();
+        List<Machine> machines = machineServiceImpl.findAllMachines();
         modelMap.addAttribute("machines", machines);
         LOGGER.info("listView(...)" + machines);
         return "list-machine.html";
@@ -39,9 +39,9 @@ public class WebMachineController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws Exception {
         LOGGER.info("updateView()");
-        Machine machine = machineService.read(id);
+        Machine machine = machineServiceImpl.findMachineById(id);
         modelMap.addAttribute("machine", machine);
-        List<Department> departments = departmentService.list();
+        List<Department> departments = departmentService.findAllDepartments();
         modelMap.addAttribute("departments", departments);
         return "update-machine.html";
     }
@@ -50,7 +50,7 @@ public class WebMachineController {
     public String update(
             @ModelAttribute(name = "machine") Machine machine) {
         LOGGER.info("update()" + machine);
-        Machine savedMachine = machineService.update(machine);
+        Machine savedMachine = machineServiceImpl.updateMachine(machine);
         LOGGER.info("update(...)" + savedMachine);
         return "redirect:/machines";
     }
@@ -59,7 +59,7 @@ public class WebMachineController {
     public String createView(ModelMap modelMap) {
         LOGGER.info("createView()");
         modelMap.addAttribute("machine", new Machine());
-        List<Department> departments = departmentService.list();
+        List<Department> departments = departmentService.findAllDepartments();
         modelMap.addAttribute("departments", departments);
         return "create-machine.html";
     }
@@ -70,7 +70,7 @@ public class WebMachineController {
             @ModelAttribute(name = "Machine") Machine machine) {
         LOGGER.info("create(" + departmentId + ")");
         LOGGER.info("create(" + machine + ")");
-        Machine savedMachine = machineService.create(machine);
+        Machine savedMachine = machineServiceImpl.createMachine(machine);
         LOGGER.info("create(...)" + savedMachine);
         return "redirect:/machines";
     }
@@ -80,7 +80,7 @@ public class WebMachineController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws Exception {
         LOGGER.info("read(" + id + ")");
-        Machine machine = machineService.read(id);
+        Machine machine = machineServiceImpl.findMachineById(id);
         modelMap.addAttribute("machine", machine);
         return "read-machine.html";
     }
@@ -89,7 +89,7 @@ public class WebMachineController {
     public String delete(
             @PathVariable(name = "id") Long id) {
         LOGGER.info("delete()");
-        machineService.delete(id);
+        machineServiceImpl.deleteMachine(id);
         return "redirect:/machines";
     }
 }
