+++ src/main/java/pl/medos/cmmsApi/controllers/WebResourceController.java
@@ -7,8 +7,8 @@ import pl.medos.cmmsApi.exception.ResourceNotFoundException;
 import pl.medos.cmmsApi.model.Invoice;
 import pl.medos.cmmsApi.model.Resource;
 import pl.medos.cmmsApi.repository.InvoiceRepository;
-import pl.medos.cmmsApi.service.InvoiceService;
-import pl.medos.cmmsApi.service.ResourceService;
+import pl.medos.cmmsApi.service.impl.InvoiceServiceImpl;
+import pl.medos.cmmsApi.service.impl.ResourceServiceImpl;
 
 import java.util.List;
 import java.util.logging.Logger;
@@ -19,24 +19,24 @@ public class WebResourceController {
 
     private static final Logger LOGGER = Logger.getLogger(WebResourceController.class.getName());
 
-    private ResourceService resourceService;
-    private InvoiceService invoiceService;
+    private ResourceServiceImpl resourceServiceImpl;
+    private InvoiceServiceImpl invoiceServiceImpl;
     private final InvoiceRepository invoiceRepository;
 
 
-    public WebResourceController(ResourceService resourceService, InvoiceService invoiceService,
+    public WebResourceController(ResourceServiceImpl resourceServiceImpl, InvoiceServiceImpl invoiceServiceImpl,
                                  InvoiceRepository invoiceRepository) {
-        this.resourceService = resourceService;
-        this.invoiceService = invoiceService;
+        this.resourceServiceImpl = resourceServiceImpl;
+        this.invoiceServiceImpl = invoiceServiceImpl;
         this.invoiceRepository = invoiceRepository;
     }
 
     @GetMapping
     public String listView(ModelMap modelMap) {
         LOGGER.info("listView()");
-        List<Resource> resources = resourceService.list();
+        List<Resource> resources = resourceServiceImpl.findAllResources();
         modelMap.addAttribute("resources", resources);
-        List<Invoice> invoices = invoiceService.list();
+        List<Invoice> invoices = invoiceServiceImpl.findAllInvoices();
         modelMap.addAttribute("invoices", invoices);
         LOGGER.info("listView(...)");
         return "list-resource.html";
@@ -46,7 +46,7 @@ public class WebResourceController {
     public String createView(ModelMap modelMap) {
         LOGGER.info("createView()");
         modelMap.addAttribute("resource", new Resource());
-        List<Invoice> invoices = invoiceService.list();
+        List<Invoice> invoices = invoiceServiceImpl.findAllInvoices();
         modelMap.addAttribute("invoices", invoices);
         LOGGER.info("createView(...)");
         return "create-resource.html";
@@ -55,7 +55,7 @@ public class WebResourceController {
     @PostMapping(value = "/create")
     public String create(@ModelAttribute(name = "resources") Resource resource) {
         LOGGER.info("create()");
-        Resource savedResource = resourceService.create(resource);
+        Resource savedResource = resourceServiceImpl.createResource(resource);
         LOGGER.info("create(...)" + savedResource);
         return "redirect:/resources";
     }
@@ -65,7 +65,7 @@ public class WebResourceController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws ResourceNotFoundException {
         LOGGER.info("read(" + id + ")");
-        Resource resourceModel = resourceService.read(id);
+        Resource resourceModel = resourceServiceImpl.findResourceById(id);
         modelMap.addAttribute("resource", resourceModel);
         LOGGER.info("read(...)" + resourceModel);
         return "read-resource.html";
@@ -76,9 +76,9 @@ public class WebResourceController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws ResourceNotFoundException {
         LOGGER.info("update()" + id);
-        Resource resource = resourceService.read(id);
+        Resource resource = resourceServiceImpl.findResourceById(id);
         modelMap.addAttribute("resource", resource);
-        List<Invoice> invoices = invoiceService.list();
+        List<Invoice> invoices = invoiceServiceImpl.findAllInvoices();
         modelMap.addAttribute("invoices", invoices);
         LOGGER.info("update(...)");
         return "create-resource.html";
@@ -88,7 +88,7 @@ public class WebResourceController {
     public String update(
             @ModelAttribute(name = "resources") Resource resource) throws ResourceNotFoundException {
         LOGGER.info("update()" + resource);
-        Resource updatedResource = resourceService.update(resource);
+        Resource updatedResource = resourceServiceImpl.updateResource(resource);
         LOGGER.info("update(...)" + updatedResource);
         return "redirect:/resources";
     }
@@ -96,7 +96,7 @@ public class WebResourceController {
     @GetMapping(value = "/delete/{id}")
     public String delete(@PathVariable(name = "id") Long id) {
         LOGGER.info("delete(" + id + ")");
-        resourceService.delete(id);
+        resourceServiceImpl.deleteResource(id);
         LOGGER.info("delete(...)");
         return "redirect:/resources";
     }
