+++ src/main/java/pl/medos/cmmsApi/controllers/WebSupplierController.java
@@ -3,13 +3,9 @@ package pl.medos.cmmsApi.controllers;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
-import pl.medos.cmmsApi.exception.ResourceNotFoundException;
 import pl.medos.cmmsApi.exception.SupplierNotFoundException;
-import pl.medos.cmmsApi.model.Resource;
 import pl.medos.cmmsApi.model.Supplier;
-import pl.medos.cmmsApi.service.InvoiceService;
-import pl.medos.cmmsApi.service.ResourceService;
-import pl.medos.cmmsApi.service.SupplierService;
+import pl.medos.cmmsApi.service.impl.SupplierServiceImpl;
 
 import java.util.List;
 import java.util.logging.Logger;
@@ -20,16 +16,16 @@ public class WebSupplierController {
 
     private static final Logger LOGGER = Logger.getLogger(WebSupplierController.class.getName());
 
-    private SupplierService supplierService;
+    private SupplierServiceImpl supplierServiceImpl;
 
-    public WebSupplierController(SupplierService supplierService) {
-        this.supplierService = supplierService;
+    public WebSupplierController(SupplierServiceImpl supplierServiceImpl) {
+        this.supplierServiceImpl = supplierServiceImpl;
     }
 
     @GetMapping
     public String listView(ModelMap modelMap) {
         LOGGER.info("listView()");
-        List<Supplier> suppliers = supplierService.list();
+        List<Supplier> suppliers = supplierServiceImpl.list();
         modelMap.addAttribute("suppliers", suppliers);
         LOGGER.info("listView(...)");
         return "list-supplier.html";
@@ -46,7 +42,7 @@ public class WebSupplierController {
     @PostMapping(value = "/create")
     public String create(@ModelAttribute(name = "supplier") Supplier supplier) {
         LOGGER.info("create()");
-        Supplier savedSupplier = supplierService.create(supplier);
+        Supplier savedSupplier = supplierServiceImpl.create(supplier);
         LOGGER.info("create(...)" + savedSupplier);
         return "redirect:/suppliers";
     }
@@ -56,7 +52,7 @@ public class WebSupplierController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws SupplierNotFoundException {
         LOGGER.info("read(" + id + ")");
-        Supplier supplierModel = supplierService.read(id);
+        Supplier supplierModel = supplierServiceImpl.read(id);
         modelMap.addAttribute("supplier", supplierModel);
         LOGGER.info("read(...)" + supplierModel);
         return "read-supplier.html";
@@ -67,7 +63,7 @@ public class WebSupplierController {
             @PathVariable(name = "id") Long id,
             ModelMap modelMap) throws SupplierNotFoundException {
         LOGGER.info("update()" + id);
-       Supplier supplier = supplierService.read(id);
+       Supplier supplier = supplierServiceImpl.read(id);
         modelMap.addAttribute("supplier", supplier);
         LOGGER.info("update(...)");
         return "create-supplier.html";
@@ -77,7 +73,7 @@ public class WebSupplierController {
     public String update(
             @ModelAttribute(name = "supplier") Supplier supplier) throws SupplierNotFoundException {
         LOGGER.info("update()" + supplier);
-        Supplier updatedsupplier = supplierService.update(supplier);
+        Supplier updatedsupplier = supplierServiceImpl.update(supplier);
         LOGGER.info("update(...)" + updatedsupplier);
         return "redirect:/suppliers";
     }
@@ -85,7 +81,7 @@ public class WebSupplierController {
     @GetMapping(value = "/delete/{id}")
     public String delete(@PathVariable(name = "id") Long id) {
         LOGGER.info("delete(" + id + ")");
-        supplierService.delete(id);
+        supplierServiceImpl.delete(id);
         LOGGER.info("delete(...)");
         return "redirect:/suppliers";
     }
