+++ src/main/java/pl/medos/cmmsApi/repository/JobRepository.java
@@ -1,9 +1,14 @@
 package pl.medos.cmmsApi.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import pl.medos.cmmsApi.repository.entity.JobEntity;
 
-@Repository
+
 public interface JobRepository extends JpaRepository<JobEntity, Long> {
+//
+//    Optional<JobEntity> findByDepartment(String department);
+//    @Query("SELECT p from Jobs p WHERE " +
+//            " p.machine.name LIKE CONCAT('%', :query, '%')")
+//    List<JobEntity> searchJobs(String query)
 }
