+++ src/main/java/pl/medos/cmmsApi/service/EmployeeService.java
@@ -1,85 +1,18 @@
 package pl.medos.cmmsApi.service;
 
-import org.springframework.stereotype.Service;
 import pl.medos.cmmsApi.exception.EmployeeNotFoundException;
 import pl.medos.cmmsApi.model.Employee;
-import pl.medos.cmmsApi.repository.EmployeeRepository;
-import pl.medos.cmmsApi.repository.entity.EmployeeEntity;
-import pl.medos.cmmsApi.service.mapper.EmployeeMapper;
-
 import java.util.List;
-import java.util.Optional;
-import java.util.logging.Logger;
 
-@Service
-public class EmployeeService {
+public interface EmployeeService {
 
-    private static final Logger LOGGER = Logger.getLogger(EmployeeService.class.getName());
+    List<Employee> finadAllEmployees();
 
-    private EmployeeRepository employeeRepository;
-    private EmployeeMapper employeeMapper;
+    Employee createEmployee(Employee employee);
 
-    public EmployeeService(EmployeeRepository employeeRepository, EmployeeMapper employeeMapper) {
-        this.employeeRepository = employeeRepository;
-        this.employeeMapper = employeeMapper;
-    }
+    Employee findEmployeeById(Long id) throws EmployeeNotFoundException;
 
-    public List<Employee> list() {
-        LOGGER.info("list()");
-        List<EmployeeEntity> employeeEntities = employeeRepository.findAll();
-        List<Employee> employees = employeeMapper.listModels(employeeEntities);
-        LOGGER.info("list(...)");
-        return employees;
-    }
+    Employee updateEmployee(Employee employee);
 
-    public Employee create(Employee employee) {
-        LOGGER.info("create(" + employee + ")");
-        EmployeeEntity employeeEntity = employeeMapper.modelToEntity(employee);
-        EmployeeEntity savedEmployeeEntity = employeeRepository.save(employeeEntity);
-        Employee savedEmployeeModel = employeeMapper.entityToModel(savedEmployeeEntity);
-        LOGGER.info("create(...) " + savedEmployeeModel);
-        return savedEmployeeModel;
-    }
-
-    public Employee read(Long id) throws EmployeeNotFoundException {
-
-        LOGGER.info("read( " + id + " )");
-        Optional<EmployeeEntity> optionalEmployeeEntity = employeeRepository.findById(id);
-        EmployeeEntity employeeEntity = optionalEmployeeEntity.orElseThrow(
-                () -> new EmployeeNotFoundException("Brak pracownika o podanym id " + id));
-        Employee employeeModel = employeeMapper.entityToModel(employeeEntity);
-        LOGGER.info("read(...)" + employeeModel);
-        return employeeModel;
-    }
-
-    public Employee update(Employee employee) {
-
-        LOGGER.info("update( " + employee.getId() + " )");
-        EmployeeEntity employeeEntity = employeeMapper.modelToEntity(employee);
-
-        EmployeeEntity updatedEmployeeEntity = employeeRepository.save(employeeEntity);
-        Employee updatedEmployeeModel = employeeMapper.entityToModel(updatedEmployeeEntity);
-        LOGGER.info("update(...)" + updatedEmployeeModel.getId());
-        return updatedEmployeeModel;
-    }
-
-//    public Employee update(Employee employee) {
-//        LOGGER.info("update( " + employee + " )");
-//        Employee editedEmployee = employeeRepository.findById(employee.getId()).orElseThrow();
-//        editedEmployee.setId(employee.getId());
-//        editedEmployee.setFirstName(employee.getFirstName());
-//        editedEmployee.setLastName(employee.getLastName());
-//        editedEmployee.setPosition(employee.getPosition());
-//        editedEmployee.setOccupation(employee.getOccupation());
-//        Employee updatedEmployee = employeeRepository.save(editedEmployee);
-//        LOGGER.info("Employee(...)" + updatedEmployee);
-//        return updatedEmployee;
-//    }
-
-    public String delete(Long id) {
-        LOGGER.info("delete(" + id + ")");
-        employeeRepository.deleteById(id);
-        LOGGER.info("delete(...)");
-        return "Response: 200 Record " + id + "deleted!";
-    }
+    void deleteEmployee(Long id);
 }
