+++ src/main/resources/application.properties
@@ -31,6 +31,6 @@ spring.jpa.generate-ddl=true
 #spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.hibernate.ddl-auto=update
 #
-#spring.security.user.name=cris
-#spring.security.user.password=4b555041KW#01
-#spring.security.user.roles=ADMIN
+spring.security.user.name=cris
+spring.security.user.password=cris
+spring.security.user.roles=ADMIN
